update osu to use the framework's TooltipContainer
